requirements.txt
# Core packages
requests>=2.25.0
pandas>=2.0.3
python-dotenv>=1.0.0
sqlalchemy>=2.0.23
numpy>=1.20.0
scikit-learn>=1.0.0

# Reddit API
praw==7.7.1

# OpenAI API
openai>=1.6.1

# Database connections
psycopg2-binary==2.9.9

# Formatting and display
tabulate==0.9.0

# Data Lineage
networkx>=3.0.0
pyvis>=0.3.2

# MCP Framework
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.13
tiktoken>=0.5.1
